% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_converter.R
\name{conv_384_96}
\alias{conv_384_96}
\title{convert 384 to 96wp}
\usage{
conv_384_96(data_384)
}
\arguments{
\item{data_384}{long data frame where first colum is the 384 well plate and the others are information}
}
\value{
return the same format but separated as 96wp in lists.
}
\description{
convert 384 to 96wp
}
