% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_converter.R
\name{conv_96_384}
\alias{conv_96_384}
\title{covert 96wp to 384 wp}
\usage{
conv_96_384(
  p1_96 = NULL,
  p2_96 = NULL,
  p3_96 = NULL,
  p4_96 = NULL,
  gap384 = c(T, F)[2],
  plotValues = c(T, F)[1],
  isvars_factor = c(T, F)[1]
)
}
\arguments{
\item{p1_96}{data.frame with two columns, the first being the well of a 96 wp, the second value to be distributed in 384 plate}

\item{p2_96}{data.frame with two columns, the first being the well of a 96 wp, the second value to be distributed in 384 plate}

\item{p3_96}{data.frame with two columns, the first being the well of a 96 wp, the second value to be distributed in 384 plate}

\item{p4_96}{data.frame with two columns, the first being the well of a 96 wp, the second value to be distributed in 384 plate}

\item{gap384}{add gap between 384 wells}

\item{isvars_factor}{}
}
\description{
This function takes a data.frame with two columns, the first being the well of a 96 wp, the second being some values of interest, and rearranges them into a 384 plate.
}
