% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{filter_growth_outliers}
\alias{filter_growth_outliers}
\title{Function to filter outliers in growth curves}
\usage{
filter_growth_outliers(
  plate_name = NULL,
  data,
  time_control = 24,
  slope_cutoff = NULL,
  p.val_cutoff = NULL,
  intercept_sd_cutoff = NULL,
  save_diag_plots = F,
  save_plots_directory = getwd(),
  return_stats = F
)
}
\arguments{
\item{data}{data format is a long dataframe, which has the columns "Well", "Conf" and "Time"}

\item{time_control}{what time drug was added}

\item{slope_cutoff}{slope cutoff}

\item{p.val_cutoff}{pval cutoff}

\item{intercept_sd_cutoff}{the filtering will remove the mean +/- mean * sd (sd = incercept_sd_cutoff)}

\item{save_diag_plots}{T to save diagnostic plots}

\item{save_plots_directory}{dir to save diagnostic plots}

\item{return_stats}{T if should return statistics from wells}
}
\description{
returns a list including statistics and wells that passed the treshold.
}
