% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_curves.R
\name{get_growthMetrics}
\alias{get_growthMetrics}
\title{Title}
\usage{
get_growthMetrics(data, degree, interpolated_points = 10000)
}
\arguments{
\item{data}{matrix where its first column is time and second column is confluence or other metric}

\item{degree}{polynomial degree to be used when fitting}

\item{interpolated_points}{number of time points to be interpolated by regression}
}
\description{
Title
}
