% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_converter.R
\name{get_quadrant_wells}
\alias{get_quadrant_wells}
\title{get 96 wells within 384 quadrants}
\usage{
get_quadrant_wells(quadrant)
}
\arguments{
\item{quadrant}{which quadrant we want wells from}
}
\description{
get quadrant of 384 wells to identify which well is in which quadrant. Quadrants are spaced by gaps.
}
