% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_converter.R
\name{invert_plate}
\alias{invert_plate}
\title{Flip 96 well plate upside down}
\usage{
invert_plate(data, cell, plate, wells)
}
\arguments{
\item{data}{data to be corrected}

\item{cell}{cell type}

\item{plate}{plate number}

\item{wells}{wells within plate}
}
\description{
Flip 96 well plate upside down
}
