% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{plot_merge_384}
\alias{plot_merge_384}
\title{Function to plot 384 or 96 well plates}
\usage{
plot_merge_384(data, log_e = T, plot_title = NULL)
}
\arguments{
\item{data}{data should have three columns, Well, Time and Conf in long format}

\item{log_e}{log_e set to True will natural log transform confluence}

\item{plot_title}{title of plot}
}
\description{
This function plots all wells of a 384 plate grouped together in a single plot
}
