% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{plot_multi_well}
\alias{plot_multi_well}
\title{Plot as 96 or 384 binned plot}
\usage{
plot_multi_well(is_96 = c(T, F)[2], input_df, max_timepoint = NULL)
}
\arguments{
\item{is_96}{True if plate is 96, False i splate is 384}

\item{input_df}{input data frame, with tipical format from readIncuCyteData}

\item{max_timepoint}{max time point to be plotted}
}
\description{
Plot as 96 or 384 binned plot
}
