% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth_curves.R
\name{polynomial_deriv}
\alias{polynomial_deriv}
\title{functions to obtain growth rate and other metrics
get derivative from fitted curves}
\usage{
polynomial_deriv(x)
}
\arguments{
\item{x}{a regression model}
}
\description{
this function takes the first derivative, or the growth rate equation
}
