% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{Read_PlateMap}
\alias{Read_PlateMap}
\title{Read platemap metadata}
\usage{
Read_PlateMap(
  FileName,
  Number_data_levels,
  start_position = 0,
  Plate_size = c(96, 384)[1]
)
}
\arguments{
\item{FileName}{filepath}

\item{Plate_size}{}
}
\description{
Read platemap metadata
}
