% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{Read_Tecan_M200}
\alias{Read_Tecan_M200}
\title{Read Tecan M200 96wp}
\usage{
Read_Tecan_M200(
  FileName_Tecan,
  read_platemap = F,
  start_header = F,
  Plate_mapM200
)
}
\arguments{
\item{FileName_Tecan}{filepath}

\item{Plate_mapM200}{}

\item{start_headerwhich}{row is the header of the 96wp, aka <>, if select as F it will search automatically and assign.}
}
\description{
Read Tecan M200 96wp
}
