% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{Read_TecanData}
\alias{Read_TecanData}
\title{Parsing excel processed Tecan data}
\usage{
Read_TecanData(
  FileName,
  start_position = 4,
  Number_data_levels = 4,
  read_platemap = T
)
}
\arguments{
\item{FileName}{file path}

\item{start_position}{position of well A1 in Tecan file}

\item{Number_data_levels}{Number of metadata layer in platemap}

\item{read_platemap}{True refers to option to import a plate map with metadata}
}
\description{
Parsing excel processed Tecan data
}
