% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confluence_analysis.R
\name{ReadIncuCyteData}
\alias{ReadIncuCyteData}
\title{Read and parse IncuCyte Data file}
\usage{
ReadIncuCyteData(
  FileName_IncuCyte,
  read_platemap = read_platemap,
  start_header = F,
  FileName = FileName,
  Number_data_levels = Number_data_levels,
  start_position = 0,
  correct_init_seeding = T,
  skip_first_time = F,
  Plate_size,
  time_output = c("GMT", "elapsed")[2]
)
}
\arguments{
\item{FileName_IncuCyte}{path to incucyte data}

\item{read_platemap}{True if should read metadata plate}

\item{start_header}{where the data starts}

\item{FileName}{Metadata filename}

\item{Number_data_levels}{number of data levels in metadata plate}

\item{correct_init_seeding}{T if should subtract initial seeding for each well}

\item{skip_first_time}{True to skip first time point recorded}

\item{Plate_size}{384 or 96}

\item{time_output}{should function return GMT currected time points or elapsed}
}
\description{
Read and parse IncuCyte Data file
}
