% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_converter.R
\name{switch_row}
\alias{switch_row}
\title{96wp row switch}
\usage{
switch_row(data, cell, plate, row_flip1, row_flip2, cols)
}
\arguments{
\item{data}{data to be corrected}

\item{cell}{cell type}

\item{plate}{plate number}

\item{row_flip1}{row 1 to be switched with row 2}

\item{row_flip2}{row 2 to be switched with row 1}

\item{cols}{}
}
\description{
96wp row switch
}
